Feat: Implement global package
* Previously all packages were managing
the config that they need by themselves, checking if they require a
certain environment variable and throwing errors if something was not
right.

The globals package now is the single source of config for all other
packages. This will save time by reducing clutter in packages and
avoiding checking the existence of each env multiple times for each
package